Add install_deployer_ha.sh

+++ scripts/start_ldap_ha.sh
@@ -11,10 +11,24 @@ if [[ ! -d $configPath ]]; then
     mkdir -p $configPath
 fi
 
-if [[ `ip a | grep -c $myIP` -eq 1 ]]; then
-    hostname=ldap1.$ldapDomain
-else
-    hostname=ldap2.$ldapDomain
+ips=`get_infra_ips "${ldapHosts[@]}"`
+hostname=""
+hostnameIdx=1
+hosts=""
+hostAddn=""
+for ip in ${ips[@]}; do
+    if [[ `ip a | grep -c $ip` -eq 1 ]]; then
+        hostname=ldap$hostnameIdx.$ldapDomain
+    fi
+    hosts="'ldap://ldap$hostnameIdx.$ldapDomain',${hosts[@]}"
+    hostAddn="--add-host ldap$hostnameIdx.$ldapDomain:$ip ${hostAddn[@]}"
+    hostnameIdx=$((hostnameIdx+1))
+done
+hosts=${hosts:0:-1}
+
+if [[ -z $hostname ]]; then
+    echo "failed to determine hostname"
+    exit
 fi
 
 docker rm -f ldap
@@ -24,9 +38,8 @@ docker run --name ldap -d --hostname $hostname --env LDAP_REPLICATION=true \
   --env LDAP_DOMAIN=$ldapDomain \
   --env LDAP_ADMIN_PASSWORD=$ldapRootPW \
   --env LDAP_REPLICATION=true \
-  --env LDAP_REPLICATION_HOSTS="#PYTHON2BASH:['ldap://ldap1.$ldapDomain','ldap://ldap2.$ldapDomain']" \
-  --add-host ldap1.$ldapDomain:$myIP \
-  --add-host ldap2.$ldapDomain:$peerIP \
+  --env LDAP_REPLICATION_HOSTS="#PYTHON2BASH:[$hosts]" \
+  $hostAddn \
   -v $configPath:/etc/ldap/slapd.d \
   -v $dataPath:/var/lib/ldap \
   --restart 'always' \
